cleanup

+++ isrs.h
@@ -3,13 +3,15 @@
 
 void initISRs();
 
+//Wenn hier irgendwas geändert wird, MUSS ein make clean gemacht werden!
+
 /* This defines what the stack looks like after an ISR was running */
 struct regs
 {
-    unsigned int gs, fs, es, ds;      /* pushed the segs last */
-    unsigned int edi, esi, ebp, esp, ebx, edx, ecx, eax;  /* pushed by 'pusha' */
-    unsigned int int_no, err_code;    /* our 'push byte #' and ecodes do this */
-    unsigned int eip, cs, eflags, useresp, ss;   /* pushed by the processor automatically */ 
+    unsigned long gs, fs, es, ds;      /* pushed the segs last */
+    unsigned long edi, esi, ebp, esp, ebx, edx, ecx, eax;  /* pushed by 'pusha' */
+    unsigned long int_no, err_code;    /* our 'push byte #' and ecodes do this */
+    unsigned long eip, cs, eflags, useresp, ss;   /* pushed by the processor automatically */ 
 };
 
 #endif
